Fix attribute name-spaces.
Either newer version of Nokogiri is stricter on parsing attribute
namespaces, or Calais has radically changed their schema.

Either way, just about all the parts of the code that read attribute
values was broken.

This is a first pass at correcting the issues and takes it to the point
that all the tests now pass successfully.

There are probably additional issues lurking that the tests aren't
covering.  I'll fix those as DocumentCloud encounters them.