Calais (sometimes) gzip's the response
I don't think anything we're doing is triggering this.  I've verified that
the Accept-Encoding header isn't present, but Calais is still sending
gzip'ed xml as the reply.

Probably a mis-configuration on their end but is easy enough to handle.

A future enhancement might be to set the {"Accept-Encoding" => "gzip"}
header on the request, then we should get gzip data all the time.